+++ smarts/core/default_map_builder.py
@@ -18,6 +18,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
+from enum import IntEnum
 import os
 import sys
 from typing import NamedTuple, Optional, Tuple
@@ -51,36 +52,39 @@ def _clear_cache():
         gc.collect()
 
 
-_UNKNOWN_MAP = 0
-_SUMO_MAP = 1
-_OPENDRIVE_MAP = 2
-_WAYMO_MAP = 3
-_ARGOVERSE_MAP = 4
+class MapType(IntEnum):
+    """The format of a map."""
+
+    Unknown = 0
+    Sumo = 1
+    Opendrive = 2
+    Waymo = 3
+    Argoverse = 4
 
 
-def find_mapfile_in_dir(map_dir: str) -> Tuple[int, str]:
+def find_mapfile_in_dir(map_dir: str) -> Tuple[MapType, str]:
     """Looks in a given directory for a supported map file."""
     map_filename_type = {
-        "map.net.xml": _SUMO_MAP,
-        "shifted_map-AUTOGEN.net.xml": _SUMO_MAP,
-        "map.xodr": _OPENDRIVE_MAP,
+        "map.net.xml": MapType.Sumo,
+        "shifted_map-AUTOGEN.net.xml": MapType.Sumo,
+        "map.xodr": MapType.Opendrive,
     }
-    map_type = _UNKNOWN_MAP
+    map_type = MapType.Unknown
     map_path = map_dir
     for f in os.listdir(map_dir):
         cand_map_type = map_filename_type.get(f)
         if cand_map_type is not None:
             return cand_map_type, os.path.join(map_dir, f)
         if f.endswith(".net.xml"):
-            map_type = _SUMO_MAP
+            map_type = MapType.Sumo
             map_path = os.path.join(map_dir, f)
         elif f.endswith(".xodr"):
-            map_type = _OPENDRIVE_MAP
+            map_type = MapType.Opendrive
         elif ".tfrecord" in f:
-            map_type = _WAYMO_MAP
+            map_type = MapType.Waymo
             map_path = os.path.join(map_dir, f)
         elif "log_map_archive" in f:
-            map_type = _ARGOVERSE_MAP
+            map_type = MapType.Argoverse
             map_path = os.path.join(map_dir, f)
     return map_type, map_path
 
@@ -107,21 +111,21 @@ def get_road_map(map_spec) -> Tuple[Optional[RoadMap], Optional[str]]:
     if os.path.isdir(map_spec.source):
         map_type, map_source = find_mapfile_in_dir(map_spec.source)
     else:
-        map_type = _UNKNOWN_MAP
+        map_type = MapType.Unknown
         map_source = map_spec.source
         if map_source.endswith(".net.xml"):
-            map_type = _SUMO_MAP
+            map_type = MapType.Sumo
         elif map_source.endswith(".xodr"):
-            map_type = _OPENDRIVE_MAP
+            map_type = MapType.Opendrive
         elif ".tfrecord" in map_source:
-            map_type = _WAYMO_MAP
+            map_type = MapType.Waymo
 
-    if map_type == _SUMO_MAP:
+    if map_type == MapType.Sumo:
         from smarts.core.sumo_road_network import SumoRoadNetwork
 
         map_class = SumoRoadNetwork
 
-    elif map_type == _OPENDRIVE_MAP:
+    elif map_type == MapType.Opendrive:
         from smarts.core.utils.custom_exceptions import OpenDriveException
 
         try:
@@ -130,11 +134,11 @@ def get_road_map(map_spec) -> Tuple[Optional[RoadMap], Optional[str]]:
             raise OpenDriveException.required_to("use OpenDRIVE maps")
         map_class = OpenDriveRoadNetwork
 
-    elif map_type == _WAYMO_MAP:
+    elif map_type == MapType.Waymo:
         from smarts.core.waymo_map import WaymoMap
 
         map_class = WaymoMap
-    elif map_type == _ARGOVERSE_MAP:
+    elif map_type == MapType.Argoverse:
         try:
             from smarts.core.argoverse_map import (
                 ArgoverseMap,  # pytype: disable=import-error